+++ .eslintrc.js
@@ -14,6 +14,5 @@ module.exports = {
     'prettier/prettier': 'error',
     semi: ['warn', 'always'],
     curly: ['error', 'all'],
-    'no-unreachable': false,
   },
 };
