+++ package.json
@@ -151,6 +151,7 @@
     "koa-router": "^7.3.0",
     "koa-sslify": "^2.1.2",
     "koa-static": "^4.0.2",
+    "lazysizes": "^5.3.2",
     "log4js": "^6.7.1",
     "lowlight": "^3.1.0",
     "markdown-it": "^12.3.2",
