+++ pages/activity/activitySingle.pug
@@ -230,5 +230,4 @@ block scripts
   +includeJS('/external_pkgs/citySet/agency.json.js')
   +includeJS('/external_pkgs/citySet/citySet.js')
   include ../publicModules/plyr/plyr.js.pug
-  include ../publicModules/lazySizes
   include ../publicModules/imageViewer
