+++ pages/bootstrap_base.pug
@@ -68,6 +68,8 @@ block head
   +includeJS('/tools.js')
   // 前后端公用的方法 数据类型检查相关
   +includeJS("/checkData.js")
+  // 图片懒加载
+  +includeJS("/lazysizes/lazysizes.min.js")
 
   // 公式相关
   include ./publicModules/mathFormula
