+++ pages/fund/applicationForm.pug
@@ -96,5 +96,4 @@ block fundContent
 block fundScripts
   include ../publicModules/mathFormula
   include ../publicModules/plyr/plyr.js.pug
-  include ../publicModules/lazySizes
   include ../publicModules/imageViewer
