+++ pages/fund/manage/audit/audit.pug
@@ -236,7 +236,6 @@ block fundContent
 block fundScripts
   include ../../../publicModules/floatUserPanel/floatUserPanel.2.pug
   include ../../../publicModules/floatForumPanel/floatForumPanel
-  include ../../../publicModules/lazySizes
   include ../../../publicModules/plyr/plyr.js.pug
   include ../../../publicModules/imageViewer
   include ../../../publicModules/nkcSource
