+++ pages/lib/js/lazySizes.js
@@ -1,0 +1,3 @@
+export function lazySizesInit() {
+  window.lazySizes && window.lazySizes.init();
+}
