+++ pages/post/post.pug
@@ -139,7 +139,6 @@ block scripts
   include ../publicModules/post_warning
   include ../publicModules/selectDraft/selectDraft
   //include ../publicModules/plyr/plyr.js.pug
-  //include ../publicModules/lazySizes
   include ../ueditor/ueditorBase
   include ../MathJax.pug
   include ../publicModules/disabledPost/disabledPost
