--- pages/publicModules/lazySizes.pug
@@ -1,2 +1,0 @@
-if !isIncluded('lazySizes')
-  +includeJS("/external_pkgs/lazysizes/lazysizes.min.js")
