+++ pages/publicModules/moment/momentSource.pug
@@ -1,5 +1,4 @@
 if !isIncluded('publicModules/moment/momentSource.pug')
   include ../plyr/plyr.css.pug
   include ../plyr/plyr.js.pug
-  include ../lazySizes
   include ../imageViewer
