+++ pages/publicModules/nkcSource.pug
@@ -1,7 +1,6 @@
 if !isIncluded('nkcSource')
   include ../publicModules/plyr/plyr.css.pug
   include ../publicModules/plyr/plyr.js.pug
-  include ../publicModules/lazySizes
   include ../publicModules/mathFormula
   include ../publicModules/highlight/highlight
   include ../publicModules/replaceNKCUrl
