+++ pages/thread/index.pug
@@ -640,7 +640,6 @@ block scripts
   include ../publicModules/crop/crop
   include ../publicModules/moment
   include ../publicModules/hidePost/hidePost
-  //include ../publicModules/lazySizes
   include ../publicModules/moveThread/moveThread
   include ../publicModules/disabledPost/disabledPost
   include ../ueditor/dialogs/mathformulaV2/mathformula
